+++ src/handlers/index.ts
@@ -2,3 +2,4 @@ export * from './simple';
 export * from './complete';
 export * from './project';
 export * from './pick';
+export * from './path';